requirements.txt
requests[security,socks]>=2.23.0
python-dotenv>=0.12.0
python-dateutil>=2.8.1
click>=7.1.1
pyreadline>=2.1 ; platform_system == "Windows"
tabulate>=0.8.7
xlsxwriter>=1.3.1
cachetools>=4.1.1
fuzzyfinder>=2.1.0
xmltodict>=0.12.0
dataclasses ; python_version < '3.7'
marshmallow>=3.10.0
marshmallow-jsonapi>=0.24.0
dataclasses-json>=0.5.2
pyOpenSSL>=21.0.0
